requirements.txt
# Core data science and machine learning
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
torch>=2.0.0

# Data processing and visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Excel file handling
openpyxl>=3.1.0
xlrd>=2.0.0

# Model persistence
joblib>=1.3.0

# Presentation
python-pptx>=0.6.21

# Progress bars
tqdm>=4.65.0

# Optional: for development
pytest>=7.4.0  # for running tests
black>=23.0.0  # for code formatting 